Add files via upload
The Codesoft team provided me with excellent mentorship and support, allowing me to enhance my technical skills and deepen my understanding of the industry. I am grateful for the chance to have been part of such a dynamic and innovative organization, and I look forward to applying the knowledge and expertise I've gained to future endeavors.